improve type structure
