+++ src/logic/shouldRenderBasedOnError.ts
@@ -6,7 +6,7 @@ import {
   InternalFieldName,
   FieldErrors,
   FlatFieldErrors,
-  FieldNames,
+  FieldNamesMarkedBoolean,
 } from '../types';
 
 export default function shouldRenderBasedOnError<
@@ -21,8 +21,8 @@ export default function shouldRenderBasedOnError<
   errors: FieldErrors<TFieldValues>;
   error: FlatFieldErrors<TFieldValues>;
   name: InternalFieldName<TFieldValues>;
-  validFields: FieldNames<TFieldValues>;
-  fieldsWithValidation: FieldNames<TFieldValues>;
+  validFields: FieldNamesMarkedBoolean<TFieldValues>;
+  fieldsWithValidation: FieldNamesMarkedBoolean<TFieldValues>;
 }): boolean {
   const isFieldValid = isEmptyObject(error);
   const isFormValid = isEmptyObject(errors);
