+++ src/types/fieldArray.ts
@@ -1,0 +1,18 @@
+import { Control } from './form';
+import { FieldValues } from './fields';
+
+export type FieldArrayName = string;
+
+export type UseFieldArrayOptions<
+  TKeyName extends string = 'id',
+  TControl extends Control = Control
+> = {
+  name: FieldArrayName;
+  keyName?: TKeyName;
+  control?: TControl;
+};
+
+export type ArrayField<
+  TFieldArrayValues extends FieldValues = FieldValues,
+  TKeyName extends string = 'id'
+> = TFieldArrayValues & Record<TKeyName, string>;
