+++ src/types/fields.ts
@@ -1,0 +1,51 @@
+import { IsFlatObject } from './utils';
+import { ValidationRules } from './validator';
+
+export type RadioOrCheckboxOption = {
+  ref: HTMLInputElement;
+  mutationWatcher?: MutationObserver;
+};
+
+export type InternalFieldName<TFieldValues extends FieldValues> =
+  | (keyof TFieldValues & string)
+  | string;
+
+export type FieldName<TFieldValues extends FieldValues> = IsFlatObject<
+  TFieldValues
+> extends true
+  ? Extract<keyof TFieldValues, string>
+  : string;
+
+export type CustomElement<TFieldValues extends FieldValues> = {
+  name: FieldName<TFieldValues>;
+  type?: string;
+  value?: any;
+  disabled?: boolean;
+  checked?: boolean;
+  options?: HTMLOptionsCollection;
+  files?: FileList | null;
+  focus?: () => void;
+};
+
+export type FieldValue<
+  TFieldValues extends FieldValues
+> = TFieldValues[InternalFieldName<TFieldValues>];
+
+export type FieldValues = Record<string, any>;
+
+export type FieldElement<TFieldValues extends FieldValues = FieldValues> =
+  | HTMLInputElement
+  | HTMLSelectElement
+  | HTMLTextAreaElement
+  | CustomElement<TFieldValues>;
+
+export type Ref = FieldElement;
+
+export type Field = {
+  ref: Ref;
+  options?: RadioOrCheckboxOption[];
+} & ValidationRules;
+
+export type FieldRefs<TFieldValues extends FieldValues> = Partial<
+  Record<InternalFieldName<TFieldValues>, Field>
+>;
