+++ src/types/index.ts
@@ -1,3 +1,8 @@
 export * from './form';
 export * from './props';
 export * from './utils';
+export * from './resolvers';
+export * from './fields';
+export * from './fieldArray';
+export * from './errors';
+export * from './validator';
